test(cli): move output formatting test to tests dir
